+++ sidekick_core/test/flutter_command_test.dart
@@ -1,7 +1,7 @@
 import 'package:sidekick_core/sidekick_core.dart';
+import 'package:sidekick_test/sidekick_test.dart';
 import 'package:test/test.dart';
 
-import 'fake_sdk.dart';
 import 'init_test.dart';
 
 void main() {
