--- sidekick_core/test/local_testing.dart
@@ -1,8 +1,0 @@
-import 'package:dcli/dcli.dart';
-
-/// Set to true, when the code should be checked for lint warnings and code
-/// formatting
-///
-/// Usually, this should be checked only on the latest dart version, because
-/// dartfmt is updated with the sdk and may require different formatting
-final bool analyzeGeneratedCode = env['SIDEKICK_ANALYZE'] == 'true';
