+++ sidekick_test/lib/src/local_testing.dart
@@ -2,6 +2,7 @@ import 'dart:io';
 
 import 'package:dartx/dartx_io.dart';
 import 'package:dcli/dcli.dart';
+import 'package:test/test.dart';
 
 /// True when dependencies should be linked to local sidekick dependencies
 final bool shouldUseLocalDeps = env['SIDEKICK_PUB_DEPS'] != 'true';
@@ -34,3 +35,34 @@ dependency_overrides:
 /// Usually, this should be checked only on the latest dart version, because
 /// dartfmt is updated with the sdk and may require different formatting
 final bool analyzeGeneratedCode = env['SIDEKICK_ANALYZE'] == 'true';
+
+
+R insideFakeProjectWithSidekick<R>(R Function(Directory projectDir) block) {
+  final tempDir = Directory.systemTemp.createTempSync();
+  'git init ${tempDir.path}'.run;
+
+  tempDir.file('pubspec.yaml')
+    ..createSync()
+    ..writeAsStringSync('name: main_project\n');
+  tempDir.file('dash').createSync();
+
+  final fakeSidekickDir = tempDir.directory('packages/dash_sdk')
+    ..createSync(recursive: true);
+
+  fakeSidekickDir.file('pubspec.yaml')
+    ..createSync()
+    ..writeAsStringSync('name: dash_sdk\n');
+  fakeSidekickDir.directory('lib').createSync();
+
+  env['SIDEKICK_PACKAGE_HOME'] = fakeSidekickDir.absolute.path;
+
+  addTearDown(() {
+    tempDir.deleteSync(recursive: true);
+    env['SIDEKICK_PACKAGE_HOME'] = null;
+  });
+
+  return IOOverrides.runZoned(
+        () => block(tempDir),
+    getCurrentDirectory: () => tempDir,
+  );
+}
